update sql schema for shipping service

+++ sql/shipping_schema.sql
@@ -1,8 +1,8 @@
 CREATE DATABASE shipping;
 CREATE TABLE  shipping.shipments (
-    shipment_id int NOT NULL,
-    order_id int NOT NULL,
-    user_id int NOT NULL,
+    shipment_id bigint NOT NULL,
+    order_id bigint NOT NULL,
+    user_id bigint NOT NULL,
     destination varchar(255) NOT NULL,
     status varchar(255) NOT NULL,
     PRIMARY KEY (shipment_id)
